+++ examples/src/ApplicationOptionsExample/index.js
@@ -1,0 +1,20 @@
+import React, { Component } from "react";
+import { Stage } from "react-pixi-fiber";
+import Bunny from "./../Bunny";
+
+const options = {
+  backgroundColor: 0x10bb99,
+  view: document.querySelector("#id_preexisting_canvas")
+};
+
+class ApplicationOptionsExample extends Component {
+  render() {
+    return (
+      <Stage width={800} height={600} options={options}>
+        <Bunny x={400} y={300} />
+      </Stage>
+    );
+  }
+}
+
+export default ApplicationOptionsExample;